backend/secure-settings/requirements.txt
psycopg2-binary==2.9.9
cryptography==41.0.7
bcrypt==4.1.2